add with statements to sql charts

+++ client/setup-icons.js
@@ -3,7 +3,10 @@ const fs = require('fs');
 const path = require('path');
 const icons = require('@mdi/js');
 
-exports.saveMdiIcons = () => {
+let saved = false;
+const saveMdiIcons = () => {
+  if(saved) return;
+  saved = true;
   const iconsDestinationFolder = path.join(__dirname, '/static/icons');
   if(fs.existsSync(iconsDestinationFolder)) {
     fs.rm(iconsDestinationFolder, { recursive: true }, console.log);
@@ -24,5 +27,25 @@ exports.saveMdiIcons = () => {
     </svg>`;
     fs.writeFileSync(path.join(iconsDestinationFolder, `${name.slice(3)}.svg`), iconSvg, { encoding: 'utf8' });
   });
+  console.log(` Saved ${iconEntries.length} icons`);
   fs.writeFileSync(path.join(iconsDestinationFolder, "_meta.json"), JSON.stringify(iconNames, null, 2), { encoding: 'utf8' });
 }
+
+class SaveMdiIcons {
+  apply(compiler) {
+    compiler.hooks.afterEmit.tap(
+      'Save Mdi icons plugin afterEmit',
+      _stats => {
+        saveMdiIcons();
+      }
+    );
+    compiler.hooks.done.tap(
+      'Save Mdi icons plugin done',
+      _stats => {
+        saveMdiIcons();
+      }
+    );
+  }
+}
+
+module.exports = { SaveMdiIcons };
